vssource: dvd remove dgindex codepath
