+++ vssource/indexers/D2VWitch.py
@@ -5,21 +5,17 @@ from __future__ import annotations
 import re
 from fractions import Fraction
 from functools import lru_cache, partial
-from typing import TYPE_CHECKING, Callable, ClassVar, Sequence
+from typing import Callable, ClassVar
 
-from vstools import CustomValueError, SPath, core, remap_frames, vs
+from vstools import SPath, core, vs
 
 from ..dataclasses import D2VIndexFileInfo, D2VIndexFrameData, D2VIndexHeader
-from ..rff import apply_rff_array, apply_rff_video, cut_array_on_ranges
-from .base import DVDExtIndexer
-
-if TYPE_CHECKING:
-    from ..formats.dvd.parsedvd import IFOX, IFO0Title
+from .base import ExternalIndexer
 
 __all__ = ["D2VWitch"]
 
 
-class D2VWitch(DVDExtIndexer):
+class D2VWitch(ExternalIndexer):
     _bin_path = "d2vwitch"
     _ext = "d2v"
     _source_func: ClassVar[Callable[..., vs.VideoNode]] = core.lazy.d2v.Source
@@ -149,91 +145,3 @@ class D2VWitch(DVDExtIndexer):
                 )
 
         return D2VIndexFileInfo(index_path, file_idx, header, frame_data)
-
-    def parse_vts(
-        self,
-        title: IFO0Title,
-        disable_rff: bool,
-        vobidcellids_to_take: list[tuple[int, int]],
-        target_vts: IFOX,
-        output_folder: SPath,
-        vob_input_files: Sequence[SPath],
-    ) -> tuple[vs.VideoNode, list[int], list[tuple[int, int]], list[int]]:
-        dvddd = self._d2v_vobid_frameset(vob_input_files, output_folder)
-
-        if len(dvddd.keys()) == 1 and (0, 0) in dvddd:
-            raise CustomValueError(
-                "Youre indexer created a d2v file with only zeros for vobid cellid; "
-                "This usually means outdated/unpatched D2Vwitch",
-                self.parse_vts,
-            )
-
-        frameranges = [x for y in [dvddd[a] for a in vobidcellids_to_take] for x in y]
-
-        fflags, vobids, progseq = self._d2v_collect_all_frameflags(vob_input_files, output_folder)
-
-        index_file = self.index(vob_input_files, output_folder=output_folder)[0]
-        node = self._source_func(index_file, rff=False)
-
-        assert len(node) == len(fflags) == len(vobids) == len(progseq)
-
-        progseq = cut_array_on_ranges(progseq, frameranges)
-        fflags = cut_array_on_ranges(fflags, frameranges)
-        vobids = cut_array_on_ranges(vobids, frameranges)
-        node = remap_frames(node, frameranges)
-
-        rff = [(a & 1) for a in fflags]
-
-        if not disable_rff:
-            tff = [int((a & 2) >> 1) for a in fflags]
-            prog = [int((a & 0b01000000) != 0) for a in fflags]
-
-            node = apply_rff_video(node, rff, tff, prog, progseq)
-            vobids = apply_rff_array(vobids, rff, tff, progseq)
-
-        return node, rff, vobids, []
-
-    def _d2v_collect_all_frameflags(
-        self, files: Sequence[SPath], output_folder: SPath
-    ) -> tuple[list[int], list[tuple[int, int]], list[int]]:
-        index_file = self.index(files, output_folder=output_folder)[0]
-        index_info = self.get_info(index_file)
-
-        frameflagslst = list[int]()
-        vobidlst = list[tuple[int, int]]()
-        progseqlst = list[int]()
-
-        for iframe in index_info.frame_data:
-            assert isinstance(iframe, D2VIndexFrameData)
-
-            vobcell = (iframe.vob, iframe.cell)
-
-            progseq = int((iframe.info & 0b1000000000) != 0)
-
-            for a in iframe.frameflags:
-                if a != 0xFF:
-                    frameflagslst.append(a)
-                    vobidlst.append(vobcell)
-                    progseqlst.append(progseq)
-
-        return frameflagslst, vobidlst, progseqlst
-
-    def _d2v_vobid_frameset(
-        self, files: Sequence[SPath], output_folder: SPath
-    ) -> dict[tuple[int, int], list[tuple[int, int]]]:
-        _, vobids, _ = self._d2v_collect_all_frameflags(files, output_folder)
-
-        vobidset = dict[tuple[int, int], list[tuple[int, int]]]()
-        for i, a in enumerate(vobids):
-            if a not in vobidset:
-                vobidset[a] = [(i, i - 1)]
-
-            last = vobidset[a][-1]
-
-            if last[1] + 1 == i:
-                vobidset[a][-1] = (last[0], last[1] + 1)
-                continue
-
-            vobidset[a].append((i, i))
-
-        return vobidset
