+++ vssource/indexers/__init__.py
@@ -1,4 +1,3 @@
 from .base import *
 from .D2VWitch import *
-from .dvdsrc import *
 from .misc import *
